third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/arg_min_max.https.any_gpu-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [optional] argMin uint32 4D tensor, axis=1, all options
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'argMin' on 'MLGraphBuilder': Unsupported data type uint32 for argument input, must be one of [float32, float16, int32, int8, uint8]."
[FAIL] [required] argMax float32 4D tensor, axis=3, options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] [required] argMax float32 4D tensor, axis=3, options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] [required] argMax float32 4D tensor, axis=3, explicit options.outputDataType='int32'
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] [required] argMax float16 4D tensor, axis=3, options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] [required] argMax float16 4D tensor, axis=3, options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] [required] argMax float16 4D tensor, axis=3, explicit options.outputDataType='int32'
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] [optional] argMax uint32 4D tensor, axis=1, all options
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'argMax' on 'MLGraphBuilder': Unsupported data type uint32 for argument input, must be one of [float32, float16, int32, int8, uint8]."
Harness: the test ran to completion.
